user info handling towards frontend WIP
