+++ src/main/java/com/fieldcommand/payload/user/UserJson.java
@@ -1,0 +1,5 @@
+package com.fieldcommand.payload.user;
+
+public class UserJson {
+
+}
